# 1.8.2 is out! This is the biggest optimization patch