… DoW Mod Manager/SettingsManagerForm.Designer.cs
… DoW Mod Manager/SettingsManagerForm.resx
… changelog.txt