- Renamed "rl_sso_num_times_shown" to "Login Attempts" for more clarity
- Removed "Force Watch Movies" setting because it's effect was not proved. "-nomovies" is working as usual
- Optimization: changed Regex to String.Contains() in IsModRequired() method - it's 170x faster! 0.2 ms vs 34 ms
- Optimization: changed Regex to String.Contains() in GetActiveModsFromFile() method - it's 126x faster! 0.3 ms vs 38 ms
- Optimization: changed Regex to String.Contains() in GetModFoldersFromFile() method - it's 34x faster! 1 ms vs 34 ms
- Optimization: changed Regex to String.Contains() in ModIsPlayable() method - it's 10x faster now! 3 ms vs 32 ms
- Optimization: changed Directory.GetFiles() to File.Exists() in GetCurrentGameEXE() method - it's 1.45x faster now!
- Optimization: changed Directory.GetFiles() to File.Exists() in CheckForGraphicsConfigEXE() method - it's 1.45x faster now!
- Optimization: refactored StartGameWithOptions() method - it's 1.14x faster 
- Optimization: refactored GetMods() method - it's more readable and around 2% faster
- Optimization: There is no more Regex object calls in ModManager Form - it uses 35% less RAM now. 8388 Kb vs 11 300 Kb (in my case)
- Optimization: Startup time is 8% faster now - 512 ms vs 555 ms in 1.8.1 (with a lot of mods)
- Optimization: Application is a 1 Kb lighter. 1037 Kb vs 1038 Kb in 1.8.1
- Some small tweaks here and there

I have nothing against Regex. It could be useful in some situations but not in DoW Manager - all tests shows that they are slower in our cases.