@Fragjacker 2020-06-18
Oh that's some nice optimization changes right there! Guess ReGex was indeed overkill in this situation where I really only wanted to know if a word exists in a string or not. Gotta review this ASAP!