@IgorTheLight 2020-06-19
I'm glad to hear that!
DoW Mod Manager is addicting - I keep thinking "just one new feature and I
will do something else" ;-)

On Thu, Jun 18, 2020 at 8:11 PM Dennis S. <notifications@github.com> wrote:

> Oh that's some nice optimization changes right there! Guess ReGex was
> indeed overkill in this situation where I really only wanted to know if a
> word exists in a string or not. Gotta review this ASAP!
>
> —
> You are receiving this because you were assigned.
> Reply to this email directly, view it on GitHub
> <https://github.com/Fragjacker/DoW-Mod-Manager/pull/12#issuecomment-646183346>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/APM5F37O4E4XGCWAHDETPQ3RXJDDDANCNFSM4OBLUNPA>
> .
>