@Fragjacker 2020-06-19
Alright the pull request got merged. Make sure you pull from my master though because I have made some slight changes to your files, restoring my function documentation. Those comments before the functions are placed there on purpose:
```C++
/// <summary>
/// Sample Summary.
/// </summary>
/// <param name="MyParam"></param>
/// <returns></returns>
```
They are the C# anology to [JAVADOC](https://www.oracle.com/technical-resources/articles/java/javadoc-tool.html) function documentation which displays additional information when you hover over a function name in an IDE. So in the future please don't delet them 😀.