@IgorTheLight 2020-06-19
Oops!
I tried to delete the most obvious comments. Probably I went too far :-)
I will pull your version and build upon it. Also I will add the same for my
methods.

Off topic: do you have more fun with a Python or with a C#? Would you like
DoW Mod Manager to be ported to Python? Just some random thoughts - it's
not like I'm secretly working on that or something :-)
Another random thought: we have to add a license for DoW Mod Manager. I'm
sure that noone actually will steal it and start to make money off it but
still :-) Maybe something liberal like MIT, GPLv2 (maybe v3 but some people
don't like it - we have to investigate), or zlib . BSD is too liberal so
anyone could not only sell it but also make it close source so original
authors would not have new features. SONY uses that part to make Play
Station's OS closed and keep making money of it (it was originally forked
from FreeBSD)
That will ensure what could and what couldn't be done with DoW Mod Manager.


On Fri, Jun 19, 2020 at 2:20 PM Dennis S. <notifications@github.com> wrote:

> Alright the pull request got merged. Make sure you pull from my master
> though because I have made some slight changes to your files, restoring my
> function documentation. Those comments before the functions are placed
> there on purpose:
>
> /// <summary>
> /// Sample Summary.
> /// </summary>
> /// <param name="MyParam"></param>
> /// <returns></returns>
>
> They are the C# anology to JAVADOC
> <https://www.oracle.com/technical-resources/articles/java/javadoc-tool.html>
> function documentation which displays additional information when you hover
> over a function name in an IDE. So in the future please don't delet them
> 😀.
>
> —
> You are receiving this because you were assigned.
> Reply to this email directly, view it on GitHub
> <https://github.com/Fragjacker/DoW-Mod-Manager/pull/12#issuecomment-646582130>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/APM5F34MEDM4DYPEL7KTS6LRXNCYFANCNFSM4OBLUNPA>
> .
>