@Fragjacker 2020-06-20
> Oops!
I tried to delete the most obvious comments. Probably I went too far :-)
I will pull your version and build upon it. Also I will add the same for my
methods.

If you use Visual Studio 2017 or later, jsut typing three slashes and then pressing enter should auto generate those function doc stubs for you which may come in handy.

> Off topic: do you have more fun with a Python or with a C#? Would you like
DoW Mod Manager to be ported to Python? Just some random thoughts - it's
not like I'm secretly working on that or something :-)

Personally I don't judge any language albeit I have done _far_ more work so far with Python than with any other language. I'd like to get more into C/C++ since that's my most neglected language currently, which is a shame really 😁 .

> Another random thought: we have to add a license for DoW Mod Manager. I'm
sure that noone actually will steal it and start to make money off it but
still :-) Maybe something liberal like MIT, GPLv2 (maybe v3 but some people
don't like it - we have to investigate), or zlib . BSD is too liberal so
anyone could not only sell it but also make it close source so original
authors would not have new features. SONY uses that part to make Play
Station's OS closed and keep making money of it (it was originally forked
from FreeBSD)
That will ensure what could and what couldn't be done with DoW Mod Manager.

Adding a licence is indeed a good idea. I am rooting for the MIT license so far. In fact I should do this right now, better safe then sorry 😬 .